refactor: 파일 구조 변경 및 YourssuEmailSchema 네이밍 수정
