+++ src/pages/Applicants/Applicants.tsx
@@ -4,7 +4,7 @@ import {
   StyledTitle,
 } from '@/styles/pages/table.ts';
 import { useTabs } from '@yourssu/design-system-react';
-import ApplicantTab from '@/pages/Applicants/ApplicantTab/ApplicantTab.tsx';
+import ApplicantTab from '@/pages/Applicants/components/ApplicantTab/ApplicantTab.tsx';
 import { useGetApplicantStates } from '@/data/applicants/hooks/useGetApplicantStates.ts';
 import { ApplicantState } from '@/data/applicants/schema.ts';
