+++ src/pages/Members/Members.tsx
@@ -6,7 +6,7 @@ import {
 import { useTabs } from '@yourssu/design-system-react';
 import { MemberState } from '@/data/members/schema.ts';
 import { useGetMemberStates } from '@/data/members/hooks/useGetMemberStates.ts';
-import MemberTab from '@/pages/Members/MemberTab/MemberTab.tsx';
+import MemberTab from '@/pages/Members/components/MemberTab/MemberTab.tsx';
 
 export const Members = () => {
   const Tabs = useTabs<MemberState>({ defaultTab: '액티브', scrollable: true });
